requirements.txt
Django-2.2.2
pymysql-0.9.3